Replace CRT:REFRESH hack with STACK
Rather than refreshing the windows within the various draw functions,
and hackishly refreshing the screen at the start of the program in the
DRAW-INITIAL-UI function, the screen and windows are pushed onto a
stack, and that stack is refreshed after calling a draw function.

Thanks to McParen for explaining the problem and a couple of solutions,
as well as providing some examples of the solution I wound up going with
in action.